Extract CommandsCache to its own module
